all data related sources combined
