+++ tellurium/sedml/data/data.py
@@ -1,7 +1,8 @@
 """
 Testing reading data form the libsedml.
 
-Necessary to compile the latest version.
+Helper functions to work with data in SED-ML.
+This includes helper for CSV and NUML.
 
 - not able to load CompositeDescription from DataDescription
 - only 2D data, or also higher dimensional data supported?
@@ -19,12 +20,6 @@ except ImportError:
 print('Version:', libsedml.getLibSEDMLDottedVersion())
 
 import libnuml
-
-SOURCE_CSV = "./oscli.csv"
-SOURCE_NUML = "./oscli.numl"
-SOURCE_NUML_1D = "./OneDimensionalNuMLData.xml"
-SOURCE_NUML_2D = "./TwoDimensionalNuMLData.xml"
-SOURCE_NUML_2DRC = "./TwoDimensionTwoRCNuMLData.xml"
 
 
 #####################################
@@ -91,9 +86,6 @@ def parse_value(d, data=None):
         data = []
 
     print(type(d), d)
-
-
-
     if d.isCompositeValue():
         assert(isinstance(d, libnuml.CompositeValue))
         content = {
@@ -209,70 +201,129 @@ def load_csv_data(source):
     return df
 
 
-
-
-
 df_csv = load_csv_data(SOURCE_CSV)
 print(df_csv.head())
 
 exit()
 
 
+def sedml_example():
+    # --------------------------------------
+    # Sedml document
+    # --------------------------------------
 
-# --------------------------------------
-# Sedml document
-# --------------------------------------
+    # load sedml document
+    doc_sedml = libsedml.readSedMLFromFile('./data_plot_numl.xml')
+    print(doc_sedml)
 
-# load sedml document
-doc_sedml = libsedml.readSedMLFromFile('data_plot_numl.xml')
-print(doc_sedml)
+    # --------------------------------------
+    # ListOfDataDescriptions
+    # --------------------------------------
+    list_dd = doc_sedml.getListOfDataDescriptions()
+    print(list_dd)
+    print(len(list_dd))
 
-# --------------------------------------
-# ListOfDataDescriptions
-# --------------------------------------
-list_dd = doc_sedml.getListOfDataDescriptions()
-print(list_dd)
-print(len(list_dd))
+    for dd in list_dd:
+        # There can be multiple DataDescription
+        print('-' * 80)
+        print('DataDescription')
+        print('-' * 80)
 
-for dd in list_dd:
-    # There can be multiple DataDescription
-    print('-' * 80)
-    print('DataDescription')
-    print('-' * 80)
+        print('dd:', dd)
+        print('id:', dd.getId())
+        print('name:', dd.getId())
+        print('source', dd.getSource())
 
-    print('dd:', dd)
-    print('id:', dd.getId())
-    print('name:', dd.getId())
-    print('source', dd.getSource())
+        print('\n*** DataSources ***')
+        '''
+        The DataSource class extracts chunks out of the data 
+        file provided by the outer DataDescription element.
+        '''
 
-    print('\n*** DataSources ***')
-    '''
-    The DataSource class extracts chunks out of the data 
-    file provided by the outer DataDescription element.
-    '''
+        for ds in dd.getListOfDataSources():
+            print('\nds:', ds)
+            print('id:', ds.getId())
+            print('name:', ds.getName())
+            print('indexSet:', ds.getIndexSet())
+            print('slices')
+            for slice in ds.getListOfSlices():
+                print(slice)
+                print('\treference:', slice.getReference())
+                print('\tvalue:', slice.getValue())
 
-    for ds in dd.getListOfDataSources():
-        print('\nds:', ds)
-        print('id:', ds.getId())
-        print('name:', ds.getName())
-        print('indexSet:', ds.getIndexSet())
-        print('slices')
-        for slice in ds.getListOfSlices():
-            print(slice)
-            print('\treference:', slice.getReference())
-            print('\tvalue:', slice.getValue())
+        print('*** DimensionDescriptions ***')
+        '''
+        The dimensionDescription element is the data description from an NuML file.
+        '''
+        dim_description = dd.getDimensionDescription()
+        print('Dimension description:', dim_description)
 
-    print('*** DimensionDescriptions ***')
-    '''
-    The dimensionDescription element is the data description from an NuML file.
-    '''
-    dim_description = dd.getDimensionDescription()
-    print('Dimension description:', dim_description)
+        cd_top = dim_description.getCompositeDescription()
+        print(cd_top, type(cd_top))
+        # TODO: read the compositeDescription
 
-    cd_top = dim_description.getCompositeDescription()
-    print(cd_top, type(cd_top))
-    # TODO: read the compositeDescription
+
+def load_numl_data(source):
+    """ Helper function for loading data files from given source.
+
+    For more information see:
+        https://github.com/numl/numl
+
+          <dimensionDescription>
+            <compositeDescription indexType="double" name="time">
+              <compositeDescription indexType="string" name="SpeciesIds">
+                <atomicDescription valueType="double" name="Concentrations" />
+              </compositeDescription>
+            </compositeDescription>
+          </dimensionDescription>
+
+    :param source:
+    :return: data matrix
+    """
+    # FIXME: handle urn/url sources
+
+    doc_numl = libnuml.readNUMLFromFile(source)
+
+    print('source:', source)
+    print(doc_numl)
+
+    # reads all the resultComponents from the numl file
+    results = []
+
+    Nrc = doc_numl.getNumResultComponents()
+    # libnuml.NUMLDocument.getNumResultComponents()
+    print('NumResultComponents:', Nrc)
+
+    for res_comp in doc_numl.getResultComponents():
+        print(type(res_comp))
+        dd = res_comp.getDimensionDescription()
+        print(dd)
+        dimension_info = parse_dimension_description(dd)
+
+        for cvalue in dd.getComponentValues():
+            print(cvalue)
+
+    # TODO: figure out the dimensionality, and create np.array
+    # if 2D return a DataFrame (Does this work with the slicing?)
+
+    columns = None
+    data = None
+    df = None
+    return df
 
 
 if __name__ == "__main__":
-    pass
+    # -----------------------------------------------------
+    SOURCE_CSV = "./oscli.csv"
+    SOURCE_NUML = "./oscli.numl"
+    SOURCE_NUML_1D = "./OneDimensionalNuMLData.xml"
+    SOURCE_NUML_2D = "./TwoDimensionalNuMLData.xml"
+    SOURCE_NUML_2DRC = "./TwoDimensionTwoRCNuMLData.xml"
+    # -----------------------------------------------------
+
+    # load data from sources
+    df_numl = load_numl_data(SOURCE_NUML)
+    print(df_numl)
+
+    sedml_example()
+
